pyproject.toml
[tool.poetry]
name = "tree-viz-linkml"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
flask = "^3.0.0"
flask-cors = "^4.0.0"
gunicorn = "^21.2.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
